fix TulipProject::clearProject implementation: do not delete the temporary root directory as we only want to clear the project content and not delete it
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@12489 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-gui/src/TulipProject.cpp
@@ -365,7 +365,7 @@ QString TulipProject::temporaryPath() {
   return result;
 }
 
-bool TulipProject::removeAllDirPrivate(const QString &path) {
+bool TulipProject::removeAllDirPrivate(const QString &path, bool removeRootDir) {
   QFileInfo pathInfo(path);
 
   if (!pathInfo.isDir() || !pathInfo.exists())
@@ -388,12 +388,14 @@ bool TulipProject::removeAllDirPrivate(const QString &path) {
       return false;
   }
 
-  dir.rmdir(pathInfo.absoluteFilePath());
+  if (removeRootDir) {
+    dir.rmdir(pathInfo.absoluteFilePath());
+  }
   return true;
 }
 
 void TulipProject::clearProject() {
-  removeAllDirPrivate(_rootDir.absolutePath());
+  removeAllDirPrivate(_rootDir.absolutePath(), false);
   _projectFile = "";
 }
